fix: 패키지 분류 오류 수정
