+++ ParcelInvoiceMaker/ParcelInvoiceMaker/Processor/Parcel/ParcelProcessor.swift
@@ -6,19 +6,6 @@
 
 import Foundation
 
-protocol ParcelInformationPersistence {
-    
-    func save(parcelInformation: ParcelInformation)
-}
-
-struct DatabaseParcelInformationPersistence: ParcelInformationPersistence {
-    
-    func save(parcelInformation: ParcelInformation) {
-        // 데이터베이스에 주문 정보 저장
-        print("발송 정보를 데이터베이스에 저장했습니다.")
-    }
-}
-
 protocol ParcelOrderProcessor {
     
     func process(parcelInformation: ParcelInformation, onComplete: (ParcelInformation) -> Void)
